[refactoring: separate bindings]
