+++ src/tests/hashes/hashes.rs
@@ -1,9 +1,8 @@
 use base64::{alphabet, Engine};
 use base64::engine::{GeneralPurpose, GeneralPurposeConfig};
 use hashes::hex::{FromHex, ToHex};
-use crate::crypto::{UInt160, UInt512};
+use crate::crypto::{UInt160, UInt256, UInt512};
 use crate::util::base58;
-use crate::UInt256;
 
 #[test]
 fn test_base64_hash_size() {
