+++ src/tests/listdiff/llmq_verification.rs
@@ -1,13 +1,14 @@
 use std::collections::BTreeMap;
 use std::ptr::null_mut;
-use crate::lib_tests::tests::{add_insight_lookup_default, assert_diff_result, get_block_hash_by_height_default, get_llmq_snapshot_by_block_hash_default, get_masternode_list_by_block_hash_from_cache, get_merkle_root_by_hash_default, hash_destroy_default, masternode_list_destroy_default, masternode_list_save_in_cache, message_from_file, save_llmq_snapshot_default, should_process_diff_with_range_default, should_process_llmq_of_type, snapshot_destroy_default, FFIContext, get_block_height_by_hash_from_context};
-use crate::{models, process_mnlistdiff_from_message, processor_create_cache, register_processor};
+use crate::bindings::common::{processor_create_cache, register_processor};
+use crate::bindings::masternode::process_mnlistdiff_from_message;
+use crate::{models, types};
 use crate::ffi::from::FromFFI;
 use crate::ffi::to::ToFFI;
-use crate::types;
 use crate::chain::common::chain_type::ChainType;
 use crate::common::LLMQType;
-use crate::crypto::byte_util::UInt256;
+use crate::crypto::UInt256;
+use crate::lib_tests::tests::{add_insight_lookup_default, assert_diff_result, get_block_hash_by_height_default, get_llmq_snapshot_by_block_hash_default, get_masternode_list_by_block_hash_from_cache, get_merkle_root_by_hash_default, hash_destroy_default, masternode_list_destroy_default, masternode_list_save_in_cache, message_from_file, save_llmq_snapshot_default, should_process_diff_with_range_default, should_process_llmq_of_type, snapshot_destroy_default, FFIContext, get_block_height_by_hash_from_context};
 use crate::tests::block_store::init_testnet_store;
 
 #[test]
