+++ src/tests/listdiff/test_basic_bls_scheme.rs
@@ -1,9 +1,14 @@
 use std::collections::BTreeMap;
 use std::ptr::null_mut;
 use hashes::hex::FromHex;
+use crate::{models, types};
+use crate::bindings::common::{processor_create_cache, register_processor};
+use crate::bindings::masternode::process_mnlistdiff_from_message;
 use crate::chain::common::chain_type::{DevnetType, IHaveChainSettings};
 use crate::common::ChainType;
-use crate::{boxed, models, process_mnlistdiff_from_message, processor_create_cache, register_processor, ToFFI, types, UInt256};
+use crate::crypto::UInt256;
+use crate::ffi::boxer::boxed;
+use crate::ffi::to::ToFFI;
 use crate::lib_tests::tests::{add_insight_lookup_default, assert_diff_result, FFIContext, get_block_height_by_hash_from_context, get_block_hash_by_height_from_context, get_llmq_snapshot_by_block_hash_default, get_merkle_root_by_hash_default, hash_destroy_default, masternode_list_destroy_default, masternode_list_save_in_cache, MerkleBlock, message_from_file, save_llmq_snapshot_in_cache, should_process_diff_with_range_default, snapshot_destroy_default};
 use crate::tests::block_store::init_testnet_store;
